requirements.txt
numpy
pandas
boto3
python-dotenv
pillow
imagehash
tensorflow-macos
tensorflow-metal
pytest
-e .